Test materials were moved. Also modified to allow build with remaining materials. #660
