+++ parts/JavaConfig/projectName-selenium/src/test/java/xxxxxx/yyyyyy/zzzzzz/config/SeleniumContextConfig.java
@@ -1,17 +1,13 @@
 package xxxxxx.yyyyyy.zzzzzz.config;
 
 import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.firefox.FirefoxDriver;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Profile;
 import org.springframework.context.annotation.Scope;
 import org.springframework.context.support.PropertySourcesPlaceholderConfigurer;
 import org.springframework.core.io.Resource;
-
-import xxxxxx.yyyyyy.zzzzzz.selenium.webdrivers.ChromeDriverFactoryBean;
-import xxxxxx.yyyyyy.zzzzzz.selenium.webdrivers.EdgeDriverFactoryBean;
-import xxxxxx.yyyyyy.zzzzzz.selenium.webdrivers.FirefoxDriverFactoryBean;
 
 /**
  * Bean definition to SeleniumContext configure .
@@ -25,8 +21,7 @@ public class SeleniumContextConfig {
      * @return Bean of configured {@link PropertySourcesPlaceholderConfigurer}
      */
     @Bean
-    public static PropertySourcesPlaceholderConfigurer propertySourcesPlaceholderConfigurer(
-            @Value("classpath*:META-INF/spring/*.properties") Resource... properties) {
+    public static PropertySourcesPlaceholderConfigurer propertySourcesPlaceholderConfigurer(@Value("classpath*:META-INF/spring/*.properties") Resource...properties) {
         PropertySourcesPlaceholderConfigurer bean = new PropertySourcesPlaceholderConfigurer();
         bean.setLocations(properties);
         return bean;
@@ -34,35 +29,12 @@ public class SeleniumContextConfig {
 
     /**
      * Configure {@link WebDriver} bean.
-     * @return Bean of configured {@link FirefoxDriverFactoryBean}
+     * @return Bean of configured {@link FirefoxDriver}
      */
     @Bean("webDriver")
     @Scope("prototype")
-    @Profile({ "firefox", "default" })
-    public FirefoxDriverFactoryBean firefoxDriverFactoryBean() {
-        return new FirefoxDriverFactoryBean();
-    }
-
-    /**
-     * Configure {@link WebDriver} bean.
-     * @return Bean of configured {@link ChromeDriverFactoryBean}
-     */
-    @Bean("webDriver")
-    @Scope("prototype")
-    @Profile("chrome")
-    public ChromeDriverFactoryBean chromeDriverFactoryBean() {
-        return new ChromeDriverFactoryBean();
-    }
-
-    /**
-     * Configure {@link WebDriver} bean.
-     * @return Bean of configured {@link EdgeDriverFactoryBean}
-     */
-    @Bean("webDriver")
-    @Scope("prototype")
-    @Profile("edge")
-    public EdgeDriverFactoryBean edgeDriverFactoryBean() {
-        return new EdgeDriverFactoryBean();
+    public WebDriver webDriver() {
+        return new FirefoxDriver();
     }
 
 }
