+++ testScript/parts/JavaConfig/projectName-selenium/src/test/java/.gitkeep
@@ -1,0 +1,1 @@
+# you can delete me
